fcitx5-rime: package rime-data together as default schemas like ibus-rime
It provides schema settings including Pinyin and Wubi as a sane
default for users, to eliminate the necessity to write them from scratch.
This matches the behavior of ibus-rime, which is an official project
under rime.

Note that these schemas are just the default settings. If users don't
import them, or have their own settings or overrides under
`$XDG_DATA_HOME/fcitx5/rime`, user settings are always preferred over
rime-data.